Updated code and tests for PATCH /posts/:id

+++ src/actions/deletePostById.js
@@ -2,13 +2,14 @@ const db = require('../db/db')
 
 /**
  * Delete a post with the given id
- * @param {number} id 
+ * @param {number} id - id of post to delete
+ * @param {number} userId - userId making the request
  * @returns {promise} - Promise that resolves to deleted id if id existed
  *          -> promise resolves to null if id does not exist
  */
-const deletePostById = (id) => {
-  const query = `DELETE FROM posts WHERE id=$1 RETURNING id`
-  return db.oneOrNone(query, [id])
+const deletePostById = (id, userId) => {
+  const query = `DELETE FROM posts WHERE id=$1 AND user_id=$2 RETURNING id`
+  return db.oneOrNone(query, [id, userId])
 }
 
 module.exports = deletePostById
